Make the debug provider use an internal class
