+++ app_check/src/include/firebase/app_check/debug_provider.h
@@ -20,6 +20,12 @@
 namespace firebase {
 namespace app_check {
 
+/// @cond FIREBASE_APP_INTERNAL
+namespace internal {
+class DebugAppCheckProviderFactoryInternal;
+}
+/// @endcond
+
 /// Implementation of an {@link AppCheckProviderFactory} that builds
 /// DebugAppCheckProviders.
 ///
@@ -43,6 +49,10 @@ class DebugAppCheckProviderFactory : public AppCheckProviderFactory {
 
  private:
   DebugAppCheckProviderFactory();
+  
+  /// @cond FIREBASE_APP_INTERNAL
+  internal::DebugAppCheckProviderFactoryInternal* internal_;
+  /// @endcond
 };
 
 }  // namespace app_check
